collector/requirements.txt
astroid==2.15.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dill==0.3.6
duckduckgo-search==2.8.5
grpcio==1.51.3
grpcio-tools==1.51.3
idna==3.4
importlib-metadata==6.1.0
isort==5.11.5
kafka-python==2.0.2
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==3.5.0
protobuf==4.22.1
pylint==2.17.4
requests==2.28.2
tomli==2.0.1
tomlkit==0.11.8
typed-ast==1.5.4
typing-extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
zipp==3.15.0
